refactor(jangar): split material evidence trace helpers
