+++ services/jangar/src/server/supporting-primitives-stage-clearance-status.ts
@@ -1,0 +1,39 @@
+import { materialEvidenceStatusFields } from '~/server/supporting-primitives-material-evidence-trace'
+import type { StageClearanceLaunchAdmission } from '~/server/supporting-primitives-stage-clearance'
+
+export const stageClearanceStatusForStage = (admission: StageClearanceLaunchAdmission) => ({
+  mode: admission.mode,
+  admitted: admission.admitted,
+  stage: admission.stage,
+  actionClass: admission.actionClass,
+  packetId: admission.packet?.packet_id ?? null,
+  decision: admission.packet?.decision ?? null,
+  freshUntil: admission.packet?.fresh_until ?? null,
+  maxLaunches: admission.packet?.max_launches ?? null,
+  reasonCodes: admission.packet?.reason_codes ?? [],
+  requiredRepairAction: admission.packet?.required_repair_action ?? null,
+  rollbackTarget: admission.packet?.rollback_target ?? null,
+  clearanceMarketLedgerId: admission.clearanceMarket?.ledgerId ?? null,
+  clearanceMarketStageAdmissionId: admission.clearanceMarket?.stageAdmissionId ?? null,
+  clearanceMarketStageDecision: admission.clearanceMarket?.decision ?? null,
+  clearanceMarketSelectedRepairLotRef: admission.clearanceMarket?.selectedRepairLotRef ?? null,
+  stageCreditLedgerId: admission.stageCredit?.ledgerId ?? null,
+  stageCreditAccountId: admission.stageCredit?.accountId ?? null,
+  stageCreditDecision: admission.stageCredit?.decision ?? null,
+  stageCreditMode: admission.stageCredit?.mode ?? null,
+  stageCreditFreshUntil: admission.stageCredit?.freshUntil ?? null,
+  stageCreditReasonCodes: admission.stageCredit?.reasonCodes ?? [],
+  runnerSlotFutureId: admission.stageCredit?.runnerSlotFutureId ?? null,
+  runnerSlotFutureExpiresAt: admission.stageCredit?.runnerSlotFutureExpiresAt ?? null,
+  stageCreditSelectedRepairLotRef: admission.stageCredit?.selectedRepairLotRef ?? null,
+  evidencePressureLedgerId: admission.evidencePressure?.ledgerId ?? null,
+  evidencePressureDecision: admission.evidencePressure?.decision ?? null,
+  evidencePressureMode: admission.evidencePressure?.mode ?? null,
+  evidencePressureFreshUntil: admission.evidencePressure?.freshUntil ?? null,
+  evidencePressureReasonCodes: admission.evidencePressure?.reasonCodes ?? [],
+  evidencePressureWatchBackoffState: admission.evidencePressure?.watchBackoffState ?? null,
+  evidencePressureRequiredRepairReceipts: admission.evidencePressure?.requiredRepairReceipts ?? [],
+  ...materialEvidenceStatusFields(admission.materialEvidence),
+  reason: admission.reason,
+  message: admission.message,
+})
